@@ -3,8 +3,9 @@ package stocks
 import (
 	"encoding/json"
 	"fmt"
-	"unibot-tg/config"
-	"unibot-tg/utils"
+
+	"github.com/s-akhmedoff/unibot-tg/config"
+	"github.com/s-akhmedoff/unibot-tg/utils"
 )
 
 type globalQuote struct {
@@ -27,23 +28,32 @@ func (s stock) String() string {
 	if s.GlobalQuote.Symbol == "" {
 		return "Failed to get stock, try another one"
 	}
-	result := fmt.Sprintf("Symbol: %s\nOpen: %s\nHigh: %s\tLow: %s\nPrice: %s\nVolume: %s\nPrevious Close: %s\nChange: %s\tChange Percent: %s\n", s.GlobalQuote.Symbol, s.GlobalQuote.Open, s.GlobalQuote.High, s.GlobalQuote.Low, s.GlobalQuote.Price, s.GlobalQuote.Volume, s.GlobalQuote.PrevCl, s.GlobalQuote.Change, s.GlobalQuote.Percent)
+
+	result := fmt.Sprintf("Symbol: %s\nOpen: %s\nHigh: %s\tLow: %s\nPrice:"+
+		" %s\nVolume: %s\nPrevious Close: %s\nChange: %s\tChange Percent: %s\n", s.GlobalQuote.Symbol,
+		s.GlobalQuote.Open, s.GlobalQuote.High,
+		s.GlobalQuote.Low, s.GlobalQuote.Price,
+		s.GlobalQuote.Volume, s.GlobalQuote.PrevCl,
+		s.GlobalQuote.Change, s.GlobalQuote.Percent)
+
 	return result
 }
 
-//GetStocks - ...
+// GetStocks - ...
 func GetStocks(symbol string, config config.Config) string {
 	if symbol == "" {
 		utils.SetDefaultValue(&symbol, utils.StocksDefaultValue)
 	}
 
-	url := fmt.Sprintf(utils.StocksAPIURL, symbol, config.CurrencyKey)
+	url := fmt.Sprintf(utils.StocksAPIURL, symbol, config.Currency.APIKey)
 
 	stocks := new(stock)
 	err := json.Unmarshal(utils.GetJSON(url), stocks)
 	utils.FailOnError(err, "Failed to unmarshal data")
+
 	if err != nil {
 		return "Service is busy now!"
 	}
+
 	return stocks.String()
 }
