@@ -3,15 +3,20 @@ package generate
 import (
 	"bytes"
 	"os/exec"
-	"unibot-tg/utils"
+
+	"github.com/s-akhmedoff/unibot-tg/utils"
 )
 
-//Password - ...
+// Password - ...
 func Password(arg string) string {
 	cmd := exec.Command("./../../utils/generate/generator.sh", arg)
+
 	var out bytes.Buffer
+
 	cmd.Stdout = &out
+
 	err := cmd.Run()
 	utils.FailOnError(err, "Failed to execute shell script")
- 	return out.String()
+
+	return out.String()
 }
