@@ -1,34 +1,35 @@
 package utils
 
 const (
-	//WeatherAPIURL - ...
+	// WeatherAPIURL - ...
 	WeatherAPIURL = "https://api.openweathermap.org/data/2.5/weather?q=%s&units=metric&appid=%s"
-	//WeatherDefaultValue - ...
+	// WeatherDefaultValue - ...
 	WeatherDefaultValue = "Tashkent"
 
-	//NewsAPIURL - ...
+	// NewsAPIURL - ...
 	NewsAPIURL = "http://newsapi.org/v2/everything?domains=%s&pageSize=30&apiKey=%s"
-	//NewsDefaultValue - ...
+	// NewsDefaultValue - ...
 	NewsDefaultValue = "yandex.ru"
 
-	//StocksAPIURL - ...
+	// StocksAPIURL - ...
 	StocksAPIURL = "https://www.alphavantage.co/query?function=GLOBAL_QUOTE&symbol=%s&apikey=%s"
-	//StocksDefaultValue - ...
+	// StocksDefaultValue - ...
 	StocksDefaultValue = "AMD"
 
-	//DefinitionAPIURL - ...
+	// DefinitionAPIURL - ...
 	DefinitionAPIURL = "https://owlbot.info/api/v4/dictionary/%s"
-	//DefinitionDefaultValue - ...
+	// DefinitionDefaultValue - ...
 	DefinitionDefaultValue = "Word"
 
-	//CurrencyAPIURL - ...
-	CurrencyAPIURL = "https://www.alphavantage.co/query?function=CURRENCY_EXCHANGE_RATE&from_currency=%s&to_currency=%s&apikey=%s"
-	//CurrencyDefaultValueFrom - ...
+	// CurrencyAPIURL - ...
+	CurrencyAPIURL = "https://www.alphavantage.co/query?function=" +
+		"CURRENCY_EXCHANGE_RATE&from_currency=%s&to_currency=%s&apikey=%s"
+	// CurrencyDefaultValueFrom - ...
 	CurrencyDefaultValueFrom = "USD"
-	//CurrencyDefaultValueTo - ...
+	// CurrencyDefaultValueTo - ...
 	CurrencyDefaultValueTo = "UZS"
 
-	//HelpDefaultResponse - ...
+	// HelpDefaultResponse - ...
 	HelpDefaultResponse = "Supported commands:\n" +
 		"/weather city - Fetch weather in city, default: Tashkent\n" +
 		"/news domain - Show latest news from domain, default yandex.ru\n" +
